requirements.txt
# Pip install -r requirements.txt

PyAudio
SpeechRecognition
discord.py@git+https://github.com/imayhaveborkedit/discord.py#egg=discord.py
PyNaCl
wheel
